Successfully completed todolist project

+++ todolist-project/app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require('mongoose');
+const _ = require("lodash");
 // const { urlencoded } = require("body-parser");
 // const date = require(__dirname + "/date.js");
 
@@ -37,7 +38,7 @@ const defaultItems = [item1,item2,item3];
 
 const listSchema = {
     name: String,
-    item: [itemsSchema]
+    items: [itemsSchema]
 };
 
 const List = mongoose.model("List", listSchema);
@@ -82,38 +83,104 @@ app.get("/", function(req, res){
 app.post("/", function(req,res){
 
     const itemName = req.body.newItem;
+    const listName = req.body.list;
+
     const item = new Item({
         name: itemName
     });
 
-    item.save();
-    res.redirect("/");
+    if(listName == "Today"){
+        item.save();
+    res.redirect("/");    
+    }
+
+    else{
+        List.findOne({name: listName})
+        .then(function(foundList){
+            foundList.items.push(item);
+            foundList.save();
+            res.redirect("/" + listName);
+        })
+    }
+
+    
+
+    
 
 });
 
 app.get("/:customListName", function(req,res){
 
-    const customListName = req.params.customListName;
-    const list = new List({
-        name: customListName,
-        items: defaultItems
+
+    const customListName = _.capitalize(req.params.customListName);
+    List.findOne({name: customListName}).exec()
+    .then(foundList => {
+        if (!foundList) {
+            const list = new List({
+                name: customListName,
+                items: defaultItems
+            })
+         
+            list.save();
+            res.redirect("/" + customListName);
+        } else {
+            res.render("list", {listTitle: foundList.name, newListItems: foundList.items});
+        }
     })
- 
-    list.save();
+    .catch(err => {
+        console.error(err);
+    });
+
+    
+    // .then(function(foundList){
+    //     if(!foundList){
+    //         console.log("Doesn't exist");
+    //     }
+    //     else{
+    //         console.log("exist");
+    //     }
+    // })
+
+    // .catch(function(err){
+    //     console.log(err);
+    // })
+        
+    
+
+    
+    
 
 });
 
 app.post("/delete", function(req,res){
     const checkedItemId  = req.body.checkbox;
+    const listName = req.body.listName;
 
-    Item.findByIdAndRemove(checkedItemId)
-    .then(function(){
-        console.log("Success");
-        res.redirect("/");
-    })
-    .catch(function(err){
-       console.log(err); 
-    })
+    if(listName === "Today"){
+        Item.findByIdAndRemove(checkedItemId)
+        .then(function(){
+            console.log("Success");
+            res.redirect("/");
+        })
+        .catch(function(err){
+            console.log(err); 
+        })
+    }
+    else{
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}})
+        .then(function(foundList){
+            if(foundList){
+                res.redirect("/" + listName);
+            }
+        })
+        .catch(function(err){
+            if(err){
+                console.log(err);
+            }
+        })
+    }
+
+    
 
 });
 
@@ -130,3 +197,11 @@ app.post("/delete", function(req,res){
 app.listen(3000, function(){
     console.log("Server running on port 3000");
 });
+
+
+// ChatGpt
+// List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, function(err, foundList){
+//             if(!err){
+//                 res.redirect("/" + listName);
+//             }
+//         })
